Update README.md
I have added some descriptive Information about the three parts of source code. As I will rename one of it, I have put the old and the new name in it as well.
Furthermore, I have added some of the bugs and future worked planned for transparency reasons and as a reminder.